refactor: add fristLoad to state & fix code structure
