--- src/types/ReduxState.ts
@@ -1,8 +1,0 @@
-import { BooksApiResponse } from './GoogleBooks'
-
-export type State = {
-  searchInput: string
-  loading: boolean
-  error: boolean
-  fetchedBooks: BooksApiResponse[]
-}
